sms/requirements.txt
alembic==0.9.3
aliyun-python-sdk-core==2.8.5
amqp==1.4.9
aniso8601==1.2.1
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
attrs==17.3.0
Automat==0.6.0
Babel==2.1.1
backports.functools-lru-cache==1.5
billiard==3.3.0.23
bleach==1.4.2
blinker==1.4
cached-property==1.3.0
celery==3.1.25
certifi==2017.4.17
cffi==1.11.2
chardet==3.0.4
click==6.3
colorama==0.3.7
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
coverage==4.0.1
crcmod==1.7
cryptography==2.1.4
cssselect==1.0.1
decorator==4.0.4
defusedxml==0.5.0
dicttoxml==1.7.4
Django==1.11.15
django-braces==1.11.0
django-celery==3.2.1
django-contrib-comments==1.8.0
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.1.0
django-filters==0.2.1
django-rest-framework==0.1.0
django-rest-framework-docs==0.1.7
django-rest-swagger==2.2.0
django-url-filter==0.3.4
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
dominate==2.3.1
enum-compat==0.0.2
enum34==1.1.6
et-xmlfile==1.0.1
gunicorn==19.6.0
honcho==0.7.1
html5lib==0.999999999
httpie==0.9.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.16
isodate==0.5.4
itsdangerous==0.24
itypes==1.1.0
jaraco.classes==1.4.3
jaraco.collections==1.5.3
jaraco.functools==1.17
jaraco.text==1.10
jdcal==1.3
Jinja2==2.9.6
jsonpickle==0.9.4
kombu==3.0.37
limits==1.2.1
lxml==3.8.0
Mako==1.0.6
MarkupSafe==1.0
mimerender==0.6.0
mistune==0.7.4
more-itertools==4.1.0
olefile==0.45.1
oletools==0.52.3
openapi-codec==1.3.2
openpack==2.1.1
openpyxl==2.5.1
optionaldict==0.1.1
oss2==2.4.0
parse==1.8.2
parsel==1.2.0
pbr==1.8.1
Pillow==3.2.0
psycopg2==2.7.1
pyasn1==0.1.9
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.4
PyDispatcher==2.0.5
Pygments==2.0.2
PyJWT==1.4.2
pymongo==3.2.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypinyin==0.33.0
python-dateutil==2.5.3
python-editor==1.0.3
python-engineio==2.2.0
python-mimeparse==1.6.0
python-socketio==2.0.0
python-weixin==0.3.0
pytz==2017.3
qrcode==5.2.2
queuelib==1.4.2
redis==2.10.5
requests==2.9.1
requests-toolbelt==0.8.0
rq==0.5.6
rsa==3.4.2
Scrapy==1.4.0
service-identity==17.0.0
shellescape==3.4.1
simplejson==3.11.1
six==1.10.0
snownlp==0.12.3
SQLAlchemy==1.1.11
suds==0.4
Twisted==17.9.0
typing==3.6.4
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.3
visitor==0.1.3
w3lib==1.18.0
webencodings==0.5.1
wechat==0.4.16
wechatpy==1.7.5
Werkzeug==0.11.4
WTForms==2.0.2
xlrd==1.1.0
XlsXcessive==0.4.1
xmltodict==0.11.0
zeep==2.2.0
zope.interface==4.4.3
